requirements.txt
# Flask and Web Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0

#AI
google-generativeai==0.3.2

# Database
SQLAlchemy==2.0.23

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# WhatsApp/Twilio
twilio==8.11.1

# PDF Generation
reportlab==4.0.7

# Email
secure-smtplib==0.1.1

# Date/Time handling
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-flask==1.3.0

# Utilities
pydantic==2.5.2
phonenumbers==8.13.27

# Optional: n8n alternative (if not using n8n)
# langchain==0.1.0
# langchain-openai==0.0.2
